address review: privatize SingleSegmentIndexWriter::new, move compute_initial_table_size into inverted index plugin, drop load_metas and map_or
